fix: delete user with score log

+++ src/user/user.service.ts
@@ -202,7 +202,7 @@ export class UserService implements OnModuleInit {
 		for (const ps of solvedStatuses) {
 			passedCounts[String(ps.problem.difficulty)]++;
 		}
-		
+
 		return passedCounts;
 	}
 
@@ -223,7 +223,6 @@ export class UserService implements OnModuleInit {
 			repository: this.userRepository,
 			dto: { limit, page },
 		});
-
 
 		if (!!searchText) {
 			users.andWhere(
@@ -348,11 +347,14 @@ export class UserService implements OnModuleInit {
 	}
 
 	async delete(id: string): Promise<void> {
-		try {
-			await this.userRepository.delete(id);
-		} catch (error) {
-			throw new NotFoundException('User not found');
+		const user = await this.findOne({ where: { id } }, false);
+
+		if (!user) {
+			throw new NotFoundException(`User with id ${id} not found`);
 		}
+
+		await this.scoreLogRepository.delete({ user: { id } });
+		await this.userRepository.delete(id);
 	}
 
 	async uploadIcon(id: string, iconBase64: string) {
